+++ docs/CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 * #468 Fix [object Object] being sometimes shown as status [1st8]
 * #493 Roster wasn't being updated after a Roster push update [teseo, jcbrand]
 * #496 Bugfix. Pings weren't being sent out. [teseo, jcbrand]
+* #467 Fix outgoing chat messages not having a msgid when being put into sessionStorage [1st8]
 
 0.9.5 (2015-08-24)
 ------------------